fix pending quotes when there are old quotes without referenced proofs

+++ internal/mint/melting.go
@@ -71,6 +71,10 @@ func (m *Mint) CheckMeltQuoteState(quoteId string) (cashu.MeltRequestDB, error) 
 	if err != nil {
 		return quote, fmt.Errorf("m.MintDB.GetMeltRequestById(quoteId). %w", err)
 	}
+	pending_proofs, err := m.MintDB.GetProofsFromQuote(tx, quote.Quote)
+	if err != nil {
+		return quote, fmt.Errorf("m.MintDB.GetProofsFromQuote(quote.Quote). %w", err)
+	}
 
 	if quote.State == cashu.PENDING {
 
@@ -89,11 +93,6 @@ func (m *Mint) CheckMeltQuoteState(quoteId string) (cashu.MeltRequestDB, error) 
 			quote.State = cashu.PAID
 			quote.FeePaid = fee
 			quote.PaymentPreimage = preimage
-
-			pending_proofs, err := m.MintDB.GetProofsFromQuote(tx, quote.Quote)
-			if err != nil {
-				return quote, fmt.Errorf("m.MintDB.GetProofsFromQuote(quote.Quote). %w", err)
-			}
 
 			changeMessages, err := m.MintDB.GetMeltChangeByQuote(tx, quote.Quote)
 			if err != nil {
@@ -156,10 +155,6 @@ func (m *Mint) CheckMeltQuoteState(quoteId string) (cashu.MeltRequestDB, error) 
 		}
 		if status == lightning.FAILED {
 			quote.State = cashu.UNPAID
-			pending_proofs, err := m.MintDB.GetProofsFromQuote(tx, quote.Quote)
-			if err != nil {
-				return quote, fmt.Errorf("m.MintDB.GetProofsFromQuote(quote.Quote). %w", err)
-			}
 
 			err = m.MintDB.ChangeMeltRequestState(tx, quote.Quote, quote.RequestPaid, quote.State, quote.Melted, quote.FeePaid)
 			if err != nil {
@@ -170,9 +165,11 @@ func (m *Mint) CheckMeltQuoteState(quoteId string) (cashu.MeltRequestDB, error) 
 			if err != nil {
 				return quote, fmt.Errorf("m.MintDB.DeleteChangeByQuote(quote.Quote) %w", err)
 			}
-			err = m.MintDB.DeleteProofs(tx, pending_proofs)
-			if err != nil {
-				return quote, fmt.Errorf("m.MintDB.DeleteProofsByQuote(quote.Quote). %w", err)
+			if len(pending_proofs) > 0 {
+				err = m.MintDB.DeleteProofs(tx, pending_proofs)
+				if err != nil {
+					return quote, fmt.Errorf("m.MintDB.DeleteProofs(tx, pending_proofs). %w", err)
+				}
 			}
 
 		}
@@ -187,16 +184,16 @@ func (m *Mint) CheckMeltQuoteState(quoteId string) (cashu.MeltRequestDB, error) 
 }
 
 func (m *Mint) CheckPendingQuoteAndProofs(logger *slog.Logger) error {
-
 	quotes, err := m.MintDB.GetMeltQuotesByState(cashu.PENDING)
 	if err != nil {
 		return fmt.Errorf("m.MintDB.GetMeltQuotesByState(cashu.PENDING). %w", err)
 	}
 
 	for _, quote := range quotes {
+		logger.Info(fmt.Sprintf("Attempting to solve pending quote for: %v", quote))
 		quote, err := m.CheckMeltQuoteState(quote.Quote)
 		if err != nil {
-			return fmt.Errorf("m.MintDB.GetMeltQuotesByState(cashu.PENDING). %w", err)
+			return fmt.Errorf("m.CheckMeltQuoteState(quote.Quote). %w", err)
 		}
 
 		logger.Info(fmt.Sprintf("Melt quote %v state: %v", quote.Quote, quote.State))
